test: consolidate overload test files into one crate to unblock CI linking
CI on this branch failed twice with 'ld terminated with signal 7 [Bus
error]' (runner out of disk while linking) — main is green, and the
tipping factor is the three new integration-test crates, each linking
its own copy of the debug-info-heavy wfl library. Merge them into a
single tests/overload_test.rs (analyzer/typechecker/interpreter
modules), cutting the new test binaries from three to one. All 32 tests
unchanged and passing.
